Product-OrderItem one-to-many association

+++ src/main/java/com/educaweb/course/entities/Order.java
@@ -24,7 +24,7 @@ public class Order implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'", timezone = "GMT")
 	private Instant moment;
@@ -41,7 +41,7 @@ public class Order implements Serializable {
 	public Order() {
 	}
 
-	public Order(Integer id, Instant moment, OrdersStatus ordersStatus, User client) {
+	public Order(Long id, Instant moment, OrdersStatus ordersStatus, User client) {
 		this.id = id;
 		this.moment = moment;
 		this.client = client;
@@ -58,11 +58,11 @@ public class Order implements Serializable {
 		}
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
